data/model: retype all entity ids from Int to String
Books, tracks, chapters and collections keyed by a String id so a
non-numeric backend can be represented (decision-11). One migration per
database, each a create-copy-drop-rename through the tested
rebuildTableWithTextIds helper, because SQLite cannot alter a column type
or a primary key.

Scope grew past the two databases cu-71 named: Chapter and Collection
carry foreign ids into books and tracks, so retyping only BookDatabase and
TrackDatabase would have left the join columns mismatched.

Fetch2's grouping API is int throughout, and downloadGroupId is a hash, so
listeners could no longer name the book they were reporting on. The id now
travels with the request in Extras and comes back verbatim; a download
enqueued before this change has no extras and is skipped rather than
guessed, since a wrong guess marks the wrong book downloaded.

ACTIVE_TRACK (Long.MIN_VALUE + 22233L) is gone: absence of the
KEY_SEEK_TO_TRACK_WITH_ID extra carries what the magic sentinel meant.

Testing gap this closed, which matters more than the retype: verify.sh was
green while a committed migration would have crashed the app on launch.
Room validates entity against schema on open, an in-memory database is
created fresh at the current version and never migrated, and nothing
opened a database through Room. RoomSchemaTest now creates a file at the
old schema and opens it at the current one for every database. A migration
that dropped every track's parentKey - orphaning every book from its
tracks, silently, with no error - passed all 201 other tests. Each new
test is verified to bite by deliberate sabotage.

Pre-existing bugs found by touching these paths:

- an inverted, vacuous check in AudiobookMediaSessionCallback asserting a
  requested track was *not* in the track list
- activeDownloads posted LiveData before mutating, and posted the mutable
  set itself, so observers saw stale contents and the update could be
  coalesced away entirely
- a track-id sentinel passed as a millisecond time offset
- two DAO methods binding numeric parameters against a now-TEXT id, which
  matches no row silently
- two getItemId overrides parsing an id with toLong(), which throws on
  exactly the non-numeric ids this change enables

Coverage 11.96% -> 12.62%; tests 201 -> 204.

Injector removal from the model layer and the repository-interface
extraction remain undone, deferred to cu-79 and cu-80 rather than bundled
with a data-destructive migration.

Task: cu-71
Verified: ./verify.sh

+++ app/src/main/java/io/github/mattpvaughn/chronicle/data/local/ChapterRepository.kt
@@ -60,7 +60,7 @@ class ChapterRepository
             // If no chapters for this track, make a chapter from the current track
             networkChapters?.map { plexChapter ->
               plexChapter.toChapter(
-                track.id.toLong(),
+                track.id,
                 track.discNumber,
                 isAudiobookCached,
               )
